requirement.txt
pytorch==1.11.0
transformers==4.30.0
tokenizers==0.13.3
matplotlib==3.7.2
google-api-python-client==2.85.0
openai==0.27.4
nltk==3.7
spacy==3.5.3
rouge==1.0.1
pandas==1.5.1
numpy==1.23.4